Limit public header to core API
